feat: add build and test lab fields
Adds a common column with updates in the ingester. This is done because a generated column would block the database for too long on its initial update and virtual generated columns are not supported on postgres 17. An initial population of the field is done in the migration for the last 7 days of data.

Closes #1595

Closes #1596